+++ packages/editor/src/components/collaborators-overlay/use-render-cursors.ts
@@ -1,12 +1,14 @@
 import {
 	privateApis as coreDataPrivateApis,
 	SelectionType,
-	type PostEditorAwarenessState as ActiveCollaborator,
+} from '@wordpress/core-data';
+import type {
+	ResolvedSelection,
+	PostEditorAwarenessState as ActiveCollaborator,
 } from '@wordpress/core-data';
 import { useSelect } from '@wordpress/data';
 import { useEffect, useState } from '@wordpress/element';
 import { store as preferencesStore } from '@wordpress/preferences';
-import type { ResolvedSelection } from '@wordpress/core-data';
 
 import { unlock } from '../../lock-unlock';
 import { getAvatarUrl } from './get-avatar-url';