+++ test/e2e/specs/editor/various/block-bindings/custom-sources.spec.js
@@ -1,8 +1,4 @@
 /**
- * External dependencies
- */
-const path = require( 'path' );
-/**
  * WordPress dependencies
  */
 const { test, expect } = require( '@wordpress/e2e-test-utils-playwright' );
@@ -14,15 +10,12 @@ test.describe( 'Registered sources', () => {
 		await requestUtils.activatePlugin( 'gutenberg-test-block-bindings' );
 		await requestUtils.deleteAllMedia();
 		const placeholderMedia = await requestUtils.uploadMedia(
-			path.join( './test/e2e/assets', '10x10_e2e_test_image_z9T8jK.png' )
+			'./assets/10x10_e2e_test_image_z9T8jK.png'
 		);
 		imagePlaceholderSrc = placeholderMedia.source_url;
 
 		const testingImgMedia = await requestUtils.uploadMedia(
-			path.join(
-				'./test/e2e/assets',
-				'1024x768_e2e_test_image_size.jpeg'
-			)
+			'./assets/1024x768_e2e_test_image_size.jpeg'
 		);
 		testingImgSrc = testingImgMedia.source_url;
 	} );
@@ -717,10 +710,7 @@ test.describe( 'Registered sources', () => {
 			requestUtils,
 		} ) => {
 			const customFieldMedia = await requestUtils.uploadMedia(
-				path.join(
-					'./test/e2e/assets',
-					'1024x768_e2e_test_image_size.jpeg'
-				)
+				'./assets/1024x768_e2e_test_image_size.jpeg'
 			);
 			testingImgSrc = customFieldMedia.source_url;